01 - Reading List: Fixed circular dependency in booklist feature.
