+++ pruebas/01-reading-list/iivvaannxx/src/features/booklist/components/current-list.svelte
@@ -11,6 +11,8 @@
 
 <script lang='ts'>
 
+  $: console.log($currentList)
+
   $: books = $currentList.books
   $: totalBooks = (books.length === 1)
     ? '1 libro'
